# go/token: slight performance improvement for IsIdentifier